\name{bkt.mixed.sim}
\alias{bkt.mixed.sim}
\title{mixed population simulation}
\usage{
  bkt.mixed.sim(mastery.params, k.opps, n.students,
    stu.params, mastery.threshold = 0.95)
}
\arguments{
  \item{mastery.params}{}

  \item{k.opps}{}

  \item{n.students}{}

  \item{stu.params}{}

  \item{mastery.threshold}{}
}
\value{
  a simulation result like that returned by
  \code{\link{bkt.sim}}
}
\description{
  mixed population simulation
}

