\name{bkt.sim}
\alias{bkt.sim}
\title{simulate a population with a single BKT model}
\usage{
  bkt.sim(n.students, k.opps, stu.params,
    mastery.params = stu.params, mastery.threshold = 0.95)
}
\arguments{
  \item{n.students}{}

  \item{k.opps}{}

  \item{stu.params}{}

  \item{mastery.params}{}

  \item{mastery.threshold}{}
}
\value{
  a list containing the following elements: \itemize{ \item
  opps \item known \item pknown \item mastered }
}
\description{
  simulate a population with a single BKT model
}

