\name{param.dist}
\alias{param.dist}
\title{distance between two BKT parameter vectors}
\usage{
  param.dist(p1, p2)
}
\arguments{
  \item{p1}{a BKT parameter object, or a matrix of same}

  \item{p2}{another BKT parameter object, or a matrix of
  same}
}
\value{
  the euclidean distance between each pair of BKT parameter
  vectors in p1, p2
}
\description{
  Calculates the Euclidean distance between two sets of BKT
  parameters
}

