\name{random.params}
\alias{random.params}
\title{generate random BKT parameters}
\usage{
  random.params()
}
\value{
  a BKT parameter object with parameters chosen uniformly
  at random in [0,1] for the init and learn parameters, and
  in [0,0.5] for the guess and slip parameters
}
\description{
  generate random BKT parameters
}

