\name{rmse}
\alias{rmse}
\title{root mean squared error for a BKT model}
\usage{
  rmse(opps, params)
}
\arguments{
  \item{opps}{the opportunity table of observed student
  actions}

  \item{params}{a BKT parameter object}
}
\value{
  the root mean squared error for the BKT model's
  predictions on this data
}
\description{
  Calculates the root mean squared error for the BKT model
  with the given parameters on the given data.
}

