\name{sim.pad.opps}
\alias{sim.pad.opps}
\title{pad out the opportunity table with simulated responses}
\usage{
  sim.pad.opps(opps, sk.params, refit.params = TRUE)
}
\arguments{
  \item{opps}{}

  \item{sk.params}{}

  \item{refit.params}{}
}
\value{
  a simulation results object, such as that returned by
  \code{\link{bkt.sim}}
}
\description{
  pad out the opportunity table with simulated responses
}

