% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data_2000}
\alias{simulation_data_2000}
\title{Simulated dataset with 2000 students}
\format{
An object of class \code{data.frame} with 19987 rows and 4 columns.
}
\usage{
simulation_data_2000
}
\description{
Simulated dataset with 2000 students
}
\keyword{datasets}
