% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data_50}
\alias{simulation_data_50}
\title{Simulated dataset}
\format{
A data frame with 50 students:
\describe{
\item{student_id}{Student ID}
\item{skill_name}{Skill name}
\item{order_id}{response ID}
\item{correct}{Binary outcome (0/1)}
}
}
\source{
Simulated using RBKT simulation functions.
}
\usage{
simulation_data_50
}
\description{
A simulated BKT dataset containing 50 students with maximum sequence length of 10.
}
\keyword{datasets}
