% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data_500}
\alias{simulation_data_500}
\title{Simulated dataset with 500 students}
\format{
An object of class \code{data.frame} with 5020 rows and 4 columns.
}
\usage{
simulation_data_500
}
\description{
Simulated dataset with 500 students
}
\keyword{datasets}
