\name{sse}
\alias{sse}
\title{sum of squared errors of the BKT model}
\usage{
  sse(opps, params)
}
\arguments{
  \item{opps}{the opportunity table of observed student
  actions}

  \item{params}{a BKT parameter object}
}
\value{
  the sum of squared errors for the BKT model's predictions
  on this data
}
\description{
  Calculates a sum of squared errors for the BKT model with
  the given parameters on the given data.
}

