% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-bkt.R
\name{bkt}
\alias{bkt}
\title{bkt(...)}
\usage{
bkt(...)
}
\arguments{
\item{parallel}{switch of parallel caculation.}

\item{seed}{seed for random number generate.}

\item{num_fits}{fits times}
}
\value{
a bkt model object, used by other bkt functions.
}
\description{
create bkt model object with initial parameters.
}
