% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-Model.R
\name{fit}
\alias{fit}
\title{fit(...)}
\usage{
fit(.Object, data_path = NULL, data = NULL, ...)
}
\arguments{
\item{data_path}{data file path.}

\item{data}{data filter.}

\item{forgets}{switch of using forget model.}

\item{skills}{skill name filter.}
}
\value{
a bkt model object, used by other bkt functions.
}
\description{
fit bkt model.
}
